requirements.txt
blobfile==3.0.0
lmdb==1.6.2
mpi4py==4.0.3
numpy==2.3.1
Pillow==10.1.0
Pillow==11.2.1
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.4
scipy==1.16.0
setuptools==65.5.1
tensorflow==2.18.0
torch==2.4.0
tqdm==4.66.1
